backend/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
mongoengine==0.24.1
passlib==1.7.4
PyJWT==2.3.0
python-dotenv==0.20.0
requests==2.27.1
boto3==1.21.29
pymongo[srv]
